+++ node-lib/src/config_files/chainstate/mod.rs
@@ -25,12 +25,14 @@ pub struct ChainstateConfigFile {
     /// The number of maximum attempts to process a block.
     pub max_db_commit_attempts: Option<usize>,
 
+    /// Whether to use the "reckless" mode during the initial block download or bootstrapping.
+    ///
+    /// In "reckless" mode the db contents is not synced to disk on each commit, which increases
+    /// performance at the cost of a potential db corruption if the system crashes.
+    pub enable_db_reckless_mode_in_ibd: Option<bool>,
+
     /// The maximum capacity of the orphan blocks pool.
     pub max_orphan_blocks: Option<usize>,
-
-    /// When importing bootstrap file, this controls the buffer sizes (min, max)
-    /// (see bootstrap import function for more information)
-    pub min_max_bootstrap_import_buffer_sizes: Option<(usize, usize)>,
 
     /// A maximum tip age in seconds.
     ///
@@ -49,8 +51,8 @@ impl From<ChainstateConfigFile> for ChainstateConfig {
     fn from(config_file: ChainstateConfigFile) -> Self {
         let ChainstateConfigFile {
             max_db_commit_attempts,
+            enable_db_reckless_mode_in_ibd,
             max_orphan_blocks,
-            min_max_bootstrap_import_buffer_sizes,
             max_tip_age,
             enable_heavy_checks,
             allow_checkpoints_mismatch,
@@ -58,8 +60,8 @@ impl From<ChainstateConfigFile> for ChainstateConfig {
 
         ChainstateConfig {
             max_db_commit_attempts: max_db_commit_attempts.into(),
+            enable_db_reckless_mode_in_ibd,
             max_orphan_blocks: max_orphan_blocks.into(),
-            min_max_bootstrap_import_buffer_sizes: min_max_bootstrap_import_buffer_sizes.into(),
             max_tip_age: max_tip_age.map(Duration::from_secs).into(),
             enable_heavy_checks,
             allow_checkpoints_mismatch,